server/requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
typing_extensions==4.12.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-rest-framework==0.1.0
Pillow==10.2.0
